refactor: 업데이트 성공 후 창 닫는 로직 통일
